+++ modules/alias/main.tf
@@ -1,5 +1,5 @@
 locals {
-  version    = element(concat(data.aws_lambda_alias.existing.*.function_version, aws_lambda_alias.with_refresh.*.function_version, aws_lambda_alias.no_refresh.*.function_version, [""]), 0)
+  version    = try(data.aws_lambda_alias.existing[0].function_version, aws_lambda_alias.with_refresh[0].function_version, aws_lambda_alias.no_refresh[0].function_version, "")
   qualifiers = zipmap(["version", "qualified_alias"], [var.create_version_async_event_config ? true : null, var.create_qualified_alias_async_event_config ? true : null])
 }
 
@@ -87,12 +87,12 @@ resource "aws_lambda_permission" "version_triggers" {
   # Error: Error adding new Lambda Permission for ... InvalidParameterValueException: We currently do not support adding policies for $LATEST.
   qualifier = local.version != "$LATEST" ? local.version : null
 
-  statement_id       = lookup(each.value, "statement_id", each.key)
-  action             = lookup(each.value, "action", "lambda:InvokeFunction")
-  principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", null)
-  source_account     = lookup(each.value, "source_account", null)
-  event_source_token = lookup(each.value, "event_source_token", null)
+  statement_id       = try(each.value.statement_id, each.key)
+  action             = try(each.value.action, "lambda:InvokeFunction")
+  principal          = try(each.value.principal, format("%s.amazonaws.com", try(each.value.service, "")))
+  source_arn         = try(each.value.source_arn, null)
+  source_account     = try(each.value.source_account, null)
+  event_source_token = try(each.value.event_source_token, null)
 }
 
 resource "aws_lambda_permission" "qualified_alias_triggers" {
@@ -101,10 +101,10 @@ resource "aws_lambda_permission" "qualified_alias_triggers" {
   function_name = var.function_name
   qualifier     = var.name
 
-  statement_id       = lookup(each.value, "statement_id", each.key)
-  action             = lookup(each.value, "action", "lambda:InvokeFunction")
-  principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", null)
-  source_account     = lookup(each.value, "source_account", null)
-  event_source_token = lookup(each.value, "event_source_token", null)
+  statement_id       = try(each.value.statement_id, each.key)
+  action             = try(each.value.action, "lambda:InvokeFunction")
+  principal          = try(each.value.principal, format("%s.amazonaws.com", try(each.value.service, "")))
+  source_arn         = try(each.value.source_arn, null)
+  source_account     = try(each.value.source_account, null)
+  event_source_token = try(each.value.event_source_token, null)
 }