+++ modules/docker-build/main.tf
@@ -2,21 +2,11 @@ data "aws_region" "current" {}
 
 data "aws_caller_identity" "this" {}
 
-data "aws_ecr_authorization_token" "token" {}
-
 locals {
   ecr_address    = coalesce(var.ecr_address, format("%v.dkr.ecr.%v.amazonaws.com", data.aws_caller_identity.this.account_id, data.aws_region.current.name))
   ecr_repo       = var.create_ecr_repo ? aws_ecr_repository.this[0].id : var.ecr_repo
   image_tag      = coalesce(var.image_tag, formatdate("YYYYMMDDhhmmss", timestamp()))
   ecr_image_name = format("%v/%v:%v", local.ecr_address, local.ecr_repo, local.image_tag)
-}
-
-provider "docker" {
-  registry_auth {
-    address  = local.ecr_address
-    username = data.aws_ecr_authorization_token.token.user_name
-    password = data.aws_ecr_authorization_token.token.password
-  }
 }
 
 resource "docker_registry_image" "this" {
@@ -27,6 +17,8 @@ resource "docker_registry_image" "this" {
     dockerfile = var.docker_file_path
     build_args = var.build_args
   }
+
+  keep_remotely = var.keep_remotely
 }
 
 resource "aws_ecr_repository" "this" {
@@ -40,4 +32,11 @@ resource "aws_ecr_repository" "this" {
   }
 
   tags = var.ecr_repo_tags
+}
+
+resource "aws_ecr_lifecycle_policy" "this" {
+  count = var.ecr_repo_lifecycle_policy != null ? 1 : 0
+
+  policy     = var.ecr_repo_lifecycle_policy
+  repository = local.ecr_repo
 }