+++ modules/docker-build/variables.tf
@@ -58,3 +58,15 @@ variable "build_args" {
   type        = map(string)
   default     = {}
 }
+
+variable "ecr_repo_lifecycle_policy" {
+  description = "A JSON formatted ECR lifecycle policy to automate the cleaning up of unused images."
+  type        = string
+  default     = null
+}
+
+variable "keep_remotely" {
+  description = "Whether to keep Docker image in the remote registry on destroy operation."
+  type        = bool
+  default     = false
+}